refactor(file structure): move feat-background/* to core/background/*
