+++ manifest.json
@@ -1,7 +1,7 @@
 {
   "background": {
     "persistent": true,
-    "page": "feat-background/index.html"
+    "page": "core/background/index.html"
   },
   "browser_action": {
     "default_icon": "assets/icons/logo-orange32x32.png",
